+++ docs/release-notes.rst
@@ -19,6 +19,9 @@ Fixed bugs
   by adding a trailing slash (issue #90)
 - fixed ``webpage.plainText``: content of invisible and script elements should not be return (issue #97)
   A flag ``plainTextAllContent`` in settings allow to have the previous behavior.
+- The webserver supports now asynchronous responses. **warning**: be sure to call the ``close()`` method
+  on the response when you finished to write data, **in any case** (synchronous or asynchronous
+  response processing) (issue #91)
 
 version 0.8.3
 =============